requirements.txt
# Requisitos básicos
numpy>=1.21.0
pandas>=1.3.0
pytest>=7.0.0
python-dotenv>=1.0.0
websockets>=10.0
yfinance>=0.1.63
ccxt>=2.0.0
tqdm>=4.65.0
loguru>=0.7.0
psutil>=5.8.0
gputil>=1.4.0

# Deep Learning e IA
torch>=2.0.0
transformers>=4.30.0
openai>=1.0.0
tensorflow>=2.13.0

# Análise Técnica
ta-lib>=0.4.24
pandas-ta>=0.3.14b0
statsmodels>=0.12.2

# Machine Learning
scikit-learn>=1.0.0
scipy>=1.6.0
joblib>=1.0.1

# Web e APIs
aiohttp>=3.8.0
requests>=2.28.0
python-binance>=1.0.12
alpaca-trade-api>=2.3.0
finnhub-python>=2.4.0
alpha_vantage>=2.3.1

# Banco de Dados
sqlalchemy>=1.4.0
pymongo>=3.11.3
redis>=3.5.3

# Desenvolvimento e testes
black>=21.5b2
flake8>=3.9.0
mypy>=0.910

# Visualização
plotly>=4.14.3
matplotlib>=3.5.0
